requirements.txt
Flask>=3.0.0
mysql-connector-python>=8.3.0
python-dotenv>=1.0.0
pymupdf>=1.24.0
python-docx>=1.1.0
pytesseract>=0.3.10
Pillow>=10.0.0
gunicorn>=21.2.0
Werkzeug>=3.0.0
